% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_raw_files.R
\name{accumulate_raw_files}
\alias{accumulate_raw_files}
\title{Accumulates data from many raw files from iGotU}
\usage{
accumulate_raw_files(directory)
}
\description{
This function accumulates data from many .csv files output by iGotU
loggers into a single dataframe named "gps.data". The function is
expecting the .csv files to be in a single
directory and to be named in the following convention:
"tagXXXXX_M.D.Y.csv" from the iGotU software. Use a 5 digit tag number,
i.e. fill in X's or 0's or something to make the tag number exactly 5
digits. The function takes one argument, which is the directory where
the files are located. Example: accumulate_raw_files("../RawData/")
}
